requirements.txt
flask
flask-cors
mysql-connector-python
gunicorn
pytz
ics
argon2-cffi
